Created tests for wrangler worker:  * updated package.json with new npm scripts and dependencies  * refactored raw response and login form  * fixed and rewrote worker.fetch unit test  * updated CONTRIBUTING.md  * addressed the feedback and lint
