+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "jest --",
     "eject": "react-scripts eject",
     "auth-worker": "wrangler dev ./worker/index",
     "auth-worker:login": "wrangler login"
@@ -23,27 +23,29 @@
   "homepage": "https://github.com/humphd/my-photohub/#readme",
   "dependencies": {
     "@fontsource/poppins": "^4.5.10",
-    "parcel": "^2.7.0",
-    "photoswipe": "^5.3.3",
+    "@octokit/rest": "^19.0.5",
+    "bootstrap": "^5.2.2",
     "eslint": "^8.27.0",
     "http-status-codes": "^2.2.0",
-    "wrangler": "^2.1.15",
     "octokit": "^2.0.10",
-    "@octokit/rest": "^19.0.5",
+    "parcel": "^2.7.0",
+    "photoswipe": "^5.3.3",
     "react": "^18.2.0",
-    "bootstrap": "^5.2.2",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
     "react-bootstrap": "^2.6.0",
+    "react-dom": "^18.2.0",
     "react-drag-drop-files": "^2.3.7",
+    "react-scripts": "5.0.1",
     "workbox-core": "^6.5.4",
     "workbox-expiration": "^6.5.4",
     "workbox-precaching": "^6.5.4",
     "workbox-routing": "^6.5.4",
-    "workbox-strategies": "^6.5.4"
+    "workbox-strategies": "^6.5.4",
+    "wrangler": "^2.1.15"
   },
   "devDependencies": {
-    "eslint": "^8.27.0"
+    "eslint": "^8.27.0",
+    "isomorphic-fetch": "^3.0.0",
+    "jest": "^27.5.1"
   },
   "browserslist": {
     "production": [
